+++ src/tests/obj-accumulator.fake.js
@@ -3,8 +3,6 @@ const sinon = require('sinon')
 function method() {}
 function getter() {}
 
-const accumulator = sinon.spy((itemName = 'item', listName = 'list') => {
-  return { method, getter }
-})
+const accumulator = sinon.fake.returns({ method, getter })
 
 module.exports = { method, getter, accumulator }